Use more appropriate fragment shader float precision
Previously the fragment shaders always used highp, but it's more
appropriate to use highp for vertex positions, mediump for texturing,
and lowp for colors.

+++ scratchpad/WebGL/webgl-scratchpad/src/shaders/color_f.glsl
@@ -1,5 +1,6 @@
 #version 300 es
-precision highp float;
+
+precision lowp float;
 
 in vec4 vertexColor;
